// frontend/src/lib/enums/enums.ts
export enum AutomatedRequirementType {
    HTTPS_PROTOCOL_CHECKER = "https_protocol_checker",
    UNAUTHORIZED_ACCESS_CHECKER = "unauthorized_access_checker",
    VULNERABILITY_CHECKER = "vulnerability_checker"
}

export enum Categories {
    ASSET_MANAGEMENT = "Asset Management",
    PROTECT = "Protect",
    DETECT = "Detect",
    RESPOND = "Respond",
    RECOVER = "Recover",
    RISK_MANAGEMENT = "Risk Management",
    SECURITY_INVESTMENTS = "Security Investments",
    PRIVACY_AND_DATA_PROTECTION = "Privacy and Data Protection",
    TRAINING_AND_AWARENESS = "Training and Awareness",
    COLLABORATION_AND_INFORMATION_SHARING = "Collaboration and Information Sharing",
}